refactor: Update index page configuration
This commit updates the index page configuration in the App.php file. The code change sets the $indexPage variable to an empty string, removing the default 'index.php' value. This modification improves the configuration by allowing the web server to handle the removal of the index file from site URIs.